Add proxy cones, for integer hull cone
A GAP proxy cone object is just like normal GAP cone object, except
that the Normaliz cone it wraps is owned by another Normaliz cone; so
when the wrapper is garbage collected, it must not delete the wrapped
Normaliz cone. Moreover, the "parent" cone must not be deleted before
the proxy cone.