Refactor test files
Because Test() does not provide a diff of the expected and actual output, it
can be be daunting to figure out what exactly is wrong if a test fails which
has very long output. This is the case when using NmzPrintConeProperties();
here, it is difficult to distinguish whether just e.g. a new property was
added, or the content of a matrix changed.

We thus avoid NmzPrintConeProperties() in the .tst files, and instead
reproduce its output in several smaller chunks which are easier to handle
within the limitations of the Test() function.