Add support for rational input matrices
For this, we implement GAPNumberToNmz for mpq_class. Then, we change the cone
constructor to always read inputs into matrices with entries in mpq_class.
While this is slightly inefficient for integral inputs, it keeps the code much
simpler, and unless the inputs are gigantic, it shouldn't matter. This is in
fact also how the Normaliz command line tool does it.