Add script to regenerate lib/cone_property_wrappers.*
We used to have a shell script to do this, but removed it at some point
because the API was fairly stable and the gains of running it didn't seem
high. It also didn't allow us to specific custom documentation for the
property wrappers. This new API, however, does allow us to do just that.

As a result, there are now a ton of new wrappers.